Implemented encoder state machine
Changed the GPIOs around to accommodate the timers needed for the motor encoder and PWM. Also switched to a state machine for the rotary encoder in place of a timer in encoder mode. This is working for now - the encoder can increase and decrease a counter when rotating.

Also decided to add a temperature sensor to the scope of the project to measure the motor's temperature. One ADC input will be needed to measure the sensor's output.